src/functions/followup_scheduler/requirements.txt
functions-framework>=3.5.0
google-cloud-firestore>=2.14.0
google-cloud-secret-manager>=2.18.1
requests>=2.31.0
flask>=2.3.0